# swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [  {
    "description" : "API GateWay",
    "url" : "https://apigtwpharma.savispain.es"
  } ],
  "info" : {
    "version" : "1.0.0",
    "title" : "API EndPoints de gestión de Delegados",
    "description" : "Descripción de los EndPoints publicados"
  },
  "security" : [ {
    "accessCode" : [ "read", "write" ]
  } ],
...
  "paths" : {
    "/api/deleg/authenticate" : {
      "post" : {
        "description" : "Login de usuario",
        "tags" : [ "authenticate" ],
        "summary" : "Autenticación de usuarios",
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "username" : {
                    "type" : "string",
                    "example" : "supervisor002@savispain.es"
                  },
                  "password" : {
                    "type" : "string",
                    "example" : "temporal"
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Acceso permitido",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "id_pos" : {
                      "type" : "integer",
                      "example" : 99999
                    },
                    "id_user" : {
                      "type" : "integer",
                      "example" : 104425
                    },
                    "status" : {
                      "type" : "string",
                      "example" : "NORMAL"
                    },
                    "name" : {
                      "type" : "string",
                      "example" : "delegado ejemplo"
                    },
                    "phone" : {
                      "type" : "string",
                      "example" : "987654321"
                    },
                    "email" : {
                      "type" : "string",
                      "example" : "supervisor002@savispain.es"
                    },
                    "roles" : {
                      "type" : "array",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "rol" : {
                            "type" : "string"
                          }
                        }
                      },
                      "example" : [ {
                        "rol" : "ROLE_SUPERVISOR"
                      }, {
                        "rol" : "ROLE_USER"
                      } ]
                    },
                    "manufacturers" : {
                      "type" : "array",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "id_manufacturer" : {
                            "type" : "integer",
                            "example" : 101
                          },
                          "manufacturer_name" : {
                            "type" : "string",
                            "example" : "L OREAL ESPAÑA, S.A."
                          }
                        }
                      }
                    },
                    "access_token" : {
                      "type" : "string",
                      "example" : "eyJhbGciOi..."
                    },
                    "refresh_token" : {
                      "type" : "string",
                      "example" : "eyJhbGciOi..."
                    },
                    "country" : {
                      "type" : "string",
                      "example" : "es"
                    }
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid request",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "message" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          },
          "403" : {
            "description" : "Acceso no autorizado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "message" : {
                      "type" : "string",
                      "example" : "Invalid credentials!"
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Error interno",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "error" : {
                      "type" : "string",
                      "example" : "SQL Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/deleg/pos/" : {
      "post" : {
        "tags" : [ "Puntos de Venta" ],
        "summary" : "Obtener lista de puntos de venta",
        "requestBody" : {
          "required" : true,
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "access_token" : {
                    "type" : "string",
                    "example" : "eyJhbGciOi..."
                  },
                  "idDelegado" : {
                    "type" : "integer",
                    "example" : 101314
                  },
                  "idPos" : {
                    "type" : "integer",
                    "example" : 101314
                  },
                  "idManufacturer" : {
                    "type" : "integer",
                    "example" : 101
                  },
                  "quicksearch" : {
                    "type" : "string",
                    "example" : "farmademo002"
                  },
                  "pendingPaymentsLast3Months" : {
                    "type" : "integer",
                    "example" : 1
                  },
                  "pendingPaymentsLastMonth" : {
                    "type" : "integer",
                    "example" : 0
                  },
                  "invalidTicketsLast3Months" : {
                    "type" : "integer",
                    "example" : 0
                  },
                  "invalidTicketsLastMonth" : {
                    "type" : "integer",
                    "example" : 0
                  },
                  "page" : {
                    "type" : "integer",
                    "example" : 0
                  }
                }
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Listado de Puntos de venta",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "pointOfSale" : {
                      "type" : "array",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "ID_POS" : {
                            "type" : "integer",
                            "example" : 101312
                          },
                          "ID_USER" : {
                            "type" : "integer",
                            "example" : 104419
                          },
                          "sUsername" : {
                            "type" : "string",
                            "example" : "deleg006@savispain.es"
                          },
                          "manufacturer_code" : {
                            "type" : "string",
                            "example" : "121212"
                          },
                          "NAME" : {
                            "type" : "string",
                            "example" : "FARMADEMO 003, S.L."
                          },
                          "CIF" : {
                            "type" : "string",
                            "example" : "B00000003"
                          },
                          "EMAIL" : {
                            "type" : "string",
                            "example" : "farmademo003@savispain.es"
                          },
                          "PHONE" : {
                            "type" : "string",
                            "example" : "987654303"
                          },
                          "CELL_PHONE" : {
                            "type" : "string",
                            "example" : null
                          },
                          "CONTACT_PERSON" : {
                            "type" : "string",
                            "example" : "Farma Demo 003"
                          },
                          "MAILING_NOTIFICATION" : {
                            "type" : "boolean",
                            "example" : true
                          },
                          "MAIN_STREET" : {
                            "type" : "string",
                            "example" : "Calle Demo 003"
                          },
                          "MAIN_CITY" : {
                            "type" : "string",
                            "example" : "ARANJUEZ"
                          },
                          "MAIN_STATE" : {
                            "type" : "string",
                            "example" : "MADRID"
                          },
                          "MAIN_ZIP_CODE" : {
                            "type" : "integer",
                            "example" : 28300
                          },
                          "POS_APP" : {
                            "type" : "boolean",
                            "example" : null
                          },
                          "CATEGORY" : {
                            "type" : "string",
                            "example" : "NORMAL"
                          },
                          "REMOTE_STORE_ID" : {
                            "type" : "string",
                            "example" : "987654303"
                          },
                          "COUNTRY" : {
                            "type" : "string",
                            "example" : "ES"
                          },
                          "PAGE" : {
                            "type" : "integer",
                            "example" : 1
                          },
                          "ROWS_OF_PAGE" : {
                            "type" : "integer",
                            "example" : 0
                          },
                          "letters" : {
                            "type" : "integer",
                            "example" : 3
                          },
                          "invalid_prizes" : {
                            "type" : "integer",
                            "example" : 4
                          },
                          "honor_date" : {
                            "type" : "string",
                            "example" : null
                          },
                          "amount" : {
                            "type" : "integer",
                            "example" : 140.31
                          }
                        }
                      }
                    },
                    "access_token" : {
                      "type" : "string",
                      "example" : "eyJhbGciOi..."
                    }
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Error interno",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "error" : {
                      "type" : "string",
                      "example" : "SQL Error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
...
    "securitySchemes" : {
      "accessCode" : {
        "type" : "oauth2",
        "flows" : {
          "authorizationCode" : {
            "authorizationUrl" : "https://apigtwpharma.savispain.es/api/authenticate",
            "tokenUrl" : "https://apigtwpharma.savispain.es/api/token",
            "scopes" : {
              "write" : "allows modifying resources",
              "read" : "allows reading resources"
            }
          }
        }
      }
    }
  }
}